+++ packages/react/src/themes/base/components/Icon/index.ts
@@ -9,6 +9,8 @@ export const icons: ThemeIcons = {
   'stardust-arrow-down': fontIcon('25BE'),
   'stardust-arrow-end': fontIcon('25B8'),
   'stardust-arrow-up': fontIcon('25B4'),
+  'stardust-pause': fontIcon('25B6'),
+  'stardust-play': fontIcon('23F8'),
 }
 
 const emptyIcon = { icon: { content: '?' } }