+++ packages/react/src/themes/teams/componentVariables.ts
@@ -22,6 +22,8 @@ export {
   default as DropdownSelectedItem,
 } from './components/Dropdown/dropdownSelectedItemVariables'
 
+export { default as Embed } from './components/Embed/embedVariables'
+
 export { default as Flex } from './components/Flex/flexVariables'
 export { default as FlexItem } from './components/Flex/flexItemVariables'
 
@@ -63,3 +65,5 @@ export { default as Text } from './components/Text/textVariables'
 export { default as TreeTitle } from './components/Tree/treeTitleVariables'
 
 export { default as Animation } from './components/Animation/animationVariables'
+
+export { default as Video } from './components/Video/videoVariables'