+++ packages/react/src/components/Dialog/Dialog.tsx
@@ -23,7 +23,10 @@ import Portal from '../Portal/Portal'
 import Ref from '../Ref/Ref'
 import Flex from '../Flex/Flex'
 
-export interface DialogProps extends UIComponentProps, ContentComponentProps, ColorComponentProps {
+export interface DialogProps
+  extends UIComponentProps,
+    ContentComponentProps<ShorthandValue>,
+    ColorComponentProps {
   /**
    * Accessibility behavior if overridden by the user.
    * @default dialogBehavior
@@ -93,7 +96,7 @@ class Dialog extends AutoControlledComponent<ReactProps<DialogProps>, DialogStat
   static propTypes = {
     ...commonPropTypes.createCommon({
       children: false,
-      content: true,
+      content: 'shorthand',
       color: true,
     }),
     actions: customPropTypes.itemShorthand,