+++ packages/react/src/themes/types.ts
@@ -326,6 +326,7 @@ export interface ThemeComponentStylesInput {
   Dropdown?: ComponentSlotStylesInput
   DropdownItem?: ComponentSlotStylesInput
   DropdownSearchInput?: ComponentSlotStylesInput
+  Embed?: ComponentSlotStylesInput
   Form?: ComponentSlotStylesInput
   FormField?: ComponentSlotStylesInput
   Grid?: ComponentSlotStylesInput
@@ -351,6 +352,10 @@ export interface ThemeComponentStylesInput {
   Segment?: ComponentSlotStylesInput
   Status?: ComponentSlotStylesInput
   Text?: ComponentSlotStylesInput
+  Tree?: ComponentSlotStylesInput
+  TreeItem?: ComponentSlotStylesInput
+  TreeTitle?: ComponentSlotStylesInput
+  Video?: ComponentSlotStylesInput
 }
 
 export interface ThemeComponentStylesPrepared {
@@ -370,6 +375,7 @@ export interface ThemeComponentStylesPrepared {
   Dropdown?: ComponentSlotStylesPrepared
   DropdownItem?: ComponentSlotStylesPrepared
   DropdownSearchInput?: ComponentSlotStylesPrepared
+  Embed?: ComponentSlotStylesPrepared
   Form?: ComponentSlotStylesPrepared
   FormField?: ComponentSlotStylesPrepared
   Grid?: ComponentSlotStylesPrepared
@@ -395,6 +401,10 @@ export interface ThemeComponentStylesPrepared {
   Segment?: ComponentSlotStylesPrepared
   Status?: ComponentSlotStylesPrepared
   Text?: ComponentSlotStylesPrepared
+  Tree?: ComponentSlotStylesPrepared
+  TreeItem?: ComponentSlotStylesPrepared
+  TreeTitle?: ComponentSlotStylesPrepared
+  Video?: ComponentSlotStylesPrepared
 }
 
 export interface ThemeComponentVariablesInput {
@@ -412,6 +422,7 @@ export interface ThemeComponentVariablesInput {
   ChatMessage?: ComponentVariablesInput
   Divider?: ComponentVariablesInput
   Dropdown?: ComponentVariablesInput
+  Embed?: ComponentVariablesInput
   Form?: ComponentVariablesInput
   FormField?: ComponentVariablesInput
   Grid?: ComponentVariablesInput
@@ -437,6 +448,10 @@ export interface ThemeComponentVariablesInput {
   Segment?: ComponentVariablesInput
   Status?: ComponentVariablesInput
   Text?: ComponentVariablesInput
+  Tree?: ComponentVariablesInput
+  TreeItem?: ComponentVariablesInput
+  TreeTitle?: ComponentVariablesInput
+  Video?: ComponentVariablesInput
 }
 
 export interface ThemeComponentVariablesPrepared {
@@ -454,6 +469,7 @@ export interface ThemeComponentVariablesPrepared {
   ChatMessage?: ComponentVariablesPrepared
   Divider?: ComponentVariablesPrepared
   Dropdown?: ComponentVariablesPrepared
+  Embed?: ComponentVariablesPrepared
   Form?: ComponentVariablesPrepared
   FormField?: ComponentVariablesPrepared
   Grid?: ComponentVariablesPrepared
@@ -479,6 +495,10 @@ export interface ThemeComponentVariablesPrepared {
   Segment?: ComponentVariablesPrepared
   Status?: ComponentVariablesPrepared
   Text?: ComponentVariablesPrepared
+  Tree?: ComponentVariablesPrepared
+  TreeItem?: ComponentVariablesPrepared
+  TreeTitle?: ComponentVariablesPrepared
+  Video?: ComponentVariablesPrepared
 }
 
 export interface Renderer extends FelaRenderer {}
@@ -529,6 +549,8 @@ export type RequiredIconNames =
   | 'stardust-arrow-end'
   | 'stardust-arrow-up'
   | 'stardust-arrow-down'
+  | 'stardust-pause'
+  | 'stardust-play'
 
 export type ThemeIcons = Partial<Record<RequiredIconNames, ThemeIconSpec>> & {
   [iconName: string]: ThemeIconSpec